mqtt-bridge/mqtt-forwarder/requirements.txt
azure-iot-device~=2.14.0
paho-mqtt
